ref(PageBanner): Move `<PageBanner>` into components/alerts folder
