+++ static/app/components/feedback/feedbackWhatsNewBanner.tsx
@@ -2,8 +2,8 @@ import {CSSProperties, useEffect} from 'react';
 
 import replaysDeadRageBackground from 'sentry-images/spot/replay-dead-rage-changelog.svg';
 
+import PageBanner from 'sentry/components/alerts/pageBanner';
 import {LinkButton} from 'sentry/components/button';
-import PageBanner from 'sentry/components/replays/pageBanner';
 import {IconBroadcast} from 'sentry/icons';
 import {t} from 'sentry/locale';
 import {trackAnalytics} from 'sentry/utils/analytics';
